Added activation of email camapign as toggle switches.

+++ src/AjaxHandler.php
@@ -25,15 +25,19 @@ class AjaxHandler
 {
     public function __construct()
     {
-        add_action('admin_init', [$this, 'act_on_option_activation_actions']);
+        add_action('admin_init', [$this, 'act_on_optin_activation_actions']);
+        add_action('admin_init', [$this, 'act_on_automation_activation_actions']);
 
         add_action('wp_ajax_mailoptin_send_test_email', array($this, 'send_test_email'));
         add_action('wp_ajax_mailoptin_create_optin_campaign', [$this, 'create_optin_campaign']);
         add_action('wp_ajax_mailoptin_create_email_campaign', [$this, 'create_email_campaign']);
         add_action('wp_ajax_mailoptin_customizer_fetch_email_list', [$this, 'customizer_fetch_email_list']);
         add_action('wp_ajax_mailoptin_optin_toggle_active', [$this, 'toggle_optin_active_status']);
+        add_action('wp_ajax_mailoptin_automation_toggle_active', [$this, 'toggle_automation_active_status']);
         add_action('wp_ajax_mailoptin_toggle_optin_activated', [$this, 'optin_listing_activated_status_toggle']);
-        add_action('wp_ajax_mailoptin_act_on_toggle_optin_activated', [$this, 'act_on_option_activation_actions']);
+        add_action('wp_ajax_mailoptin_toggle_automation_activated', [$this, 'automation_listing_activated_status_toggle']);
+        add_action('wp_ajax_mailoptin_act_on_toggle_optin_activated', [$this, 'act_on_optin_activation_actions']);
+        add_action('wp_ajax_mailoptin_act_on_toggle_automation_activated', [$this, 'act_on_automation_activation_actions']);
         add_action('wp_ajax_mailoptin_optin_type_selection', [$this, 'optin_type_selection']);
 
         add_action('wp_ajax_mailoptin_create_optin_split_test', [$this, 'create_optin_split_test']);
@@ -492,8 +496,25 @@ class AjaxHandler
         wp_die();
     }
 
-    public
-    function optin_listing_activated_status_toggle()
+    public function toggle_automation_active_status()
+    {
+        check_ajax_referer('customizer-fetch-email-list', 'security');
+
+        current_user_can('administrator') || exit;
+
+        $email_campaign_id = absint($_POST['id']);
+        $status = sanitize_text_field($_POST['status']);
+
+        if ($status == 'true') {
+            EmailCampaignRepository::activate_email_campaign($email_campaign_id);
+        } else {
+            EmailCampaignRepository::deactivate_email_campaign($email_campaign_id);
+        }
+
+        wp_die();
+    }
+
+    public function optin_listing_activated_status_toggle()
     {
         current_user_can('administrator') || exit;
 
@@ -505,7 +526,19 @@ class AjaxHandler
         wp_die();
     }
 
-    public function act_on_option_activation_actions()
+    public function automation_listing_activated_status_toggle()
+    {
+        current_user_can('administrator') || exit;
+
+        $email_campaign_id = absint($_POST['id']);
+        $status = sanitize_text_field($_POST['status']);
+
+        update_option("mo_automation_campaign_activation_task_$email_campaign_id", $status);
+
+        wp_die();
+    }
+
+    public function act_on_optin_activation_actions()
     {
         global $wpdb;
         $table = $wpdb->options;
@@ -521,6 +554,29 @@ class AjaxHandler
                     OptinCampaignsRepository::activate_campaign($optin_campaign_id);
                 } else {
                     OptinCampaignsRepository::deactivate_campaign($optin_campaign_id);
+                }
+
+                delete_option($task->option_name);
+            }
+        }
+    }
+
+    public function act_on_automation_activation_actions()
+    {
+        global $wpdb;
+        $table = $wpdb->options;
+
+        $tasks = $wpdb->get_results("SELECT option_name, option_value FROM $table WHERE option_name LIKE 'mo_automation_campaign_activation_task%'");
+
+        if (!empty($tasks)) {
+            foreach ($tasks as $task) {
+                $email_campaign_id = filter_var($task->option_name, FILTER_SANITIZE_NUMBER_INT);
+                $status = $task->option_value;
+
+                if ($status == 'true') {
+                    EmailCampaignRepository::activate_email_campaign($email_campaign_id);
+                } else {
+                    EmailCampaignRepository::deactivate_email_campaign($email_campaign_id);
                 }
 
                 delete_option($task->option_name);
@@ -553,8 +609,7 @@ class AjaxHandler
      * @param ConversionDataBuilder $conversion_data
      * @return array
      */
-    public
-    static function do_optin_conversion(ConversionDataBuilder $conversion_data)
+    public static function do_optin_conversion(ConversionDataBuilder $conversion_data)
     {
         if (!is_email($conversion_data->email)) {
             return AbstractConnect::ajax_failure(
